Review PrettyBytes typing
